Reorganise, add additional check
Moved Grosmark analysis to its own folder. Moved activity_trial_position
to prevent circular import. Added additional check to grosmark place
field analysis
