+++ viral/two_photon.py
@@ -18,17 +18,17 @@ sys.path.append(str(HERE.parent.parent))
 
 sns.set_theme(context="talk", style="ticks")
 
-from viral.imaging_utils import trial_is_imaged
+# from viral.grosmark_analysis import grosmark_place_field
+from viral.classifiers import do_classify
+from viral.grosmark_analysis import grosmark_place_field
+from viral.imaging_utils import activity_trial_position, trial_is_imaged
 from viral.rastermap_utils import get_ITI_start_frame
 from viral.utils import (
     array_bin_mean,
     degrees_to_cm,
     get_wheel_circumference_from_rig,
-    has_five_consecutive_trues,
     normalize,
     remove_consecutive_ones,
-    shuffle_rows,
-    sort_matrix_peak,
 )
 
 from viral.models import Cached2pSession, TrialInfo
@@ -67,56 +67,6 @@ def get_dff(mouse: str, date: str) -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
 
     dff = compute_dff(subtract_neuropil(f_raw, f_neu))
     return dff, spks, denoised
-
-
-def activity_trial_position(
-    trial: TrialInfo,
-    dff: np.ndarray,
-    wheel_circumference: float,
-    bin_size: int = 1,
-    start: int = 10,
-    max_position: int = 170,
-    verbose: bool = False,
-    do_shuffle: bool = False,
-) -> np.ndarray:
-
-    # TODO: remove non running epochs?
-
-    position = degrees_to_cm(
-        np.array(trial.rotary_encoder_position), wheel_circumference
-    )
-
-    frame_position = np.array(
-        [
-            state.closest_frame_start
-            for state in trial.states_info
-            if state.name
-            in ["trigger_panda", "trigger_panda_post_reward", "trigger_panda_ITI"]
-        ]
-    )
-
-    assert len(position) == len(frame_position)
-
-    dff_position = []
-
-    for bin_start in range(start, max_position, bin_size):
-        frame_idx_bin = np.unique(
-            frame_position[
-                np.logical_and(position >= bin_start, position < bin_start + bin_size)
-            ]
-        )
-        dff_bin = dff[:, frame_idx_bin]
-
-        if verbose:
-            print(f"bin_start: {bin_start}")
-            print(f"bin_end: {bin_start + bin_size}")
-            print(f"n_frames in bin: {len(frame_idx_bin)}")
-        dff_position.append(np.mean(dff_bin, axis=1))
-
-    if do_shuffle:
-        return shuffle_rows(np.array(dff_position).T)
-
-    return np.array(dff_position).T
 
 
 def remove_landmarks_from_train_matrix(train_matrix: np.ndarray) -> np.ndarray:
@@ -389,129 +339,12 @@ def place_cells_unsupervised(session: Cached2pSession, dff: np.ndarray) -> None:
     )
 
 
-def grosmark_place_field(session: Cached2pSession, spks: np.ndarray) -> np.ndarray:
-    """The position of the animal during online running epochs on the 2-m-long run belts was binned into 100,
-      2-cm spatial bins. For each cell, the as within spatial-bin firing rate was calculated across all bins
-    based on its sparsified spike estimate vector, Ssp. This firing rate by position vector was subsequently
-    smoothed with a 7.5-cm Gaussian kernel leading to the smoothed firing rate by position vector.
-    In addition, for each cell, 2,000 shuffled smoothed firing rate by position vectors were computed for each
-    cell following the per-lap randomized circular permutation of estimated activity vector, Ssp. For each cell,
-    putative PFs were defined as those in which the observed smoothed firing rate by position vectors exceeded the
-    99th percentile of their shuffled smoothed firing rate by position vectors as assessed on a per spatial-bin basis for
-    at least five consecutive spatial bins
-    """
-
-    bin_size = 2
-    start = 0
-    end = 180
-    activity_matrix = np.mean(
-        np.array(
-            [
-                activity_trial_position(
-                    trial,
-                    spks,
-                    get_wheel_circumference_from_rig("2P"),
-                    bin_size,
-                    start,
-                    end,
-                )
-                for trial in session.trials
-                if trial.texture_rewarded and trial_is_imaged(trial)
-            ]
-        ),
-        0,
-    )
-    sigma_cm = 7.5  # Desired smoothing in cm
-    sigma_bins = sigma_cm / bin_size  # Convert to bin units
-
-    # Apply Gaussian smoothing along each row (axis=1)
-    smoothed_matrix = gaussian_filter1d(activity_matrix, sigma=sigma_bins, axis=1)
-
-    shuffled_matrices = np.array(
-        [
-            gaussian_filter1d(
-                np.mean(
-                    np.array(
-                        [
-                            activity_trial_position(
-                                trial,
-                                spks,
-                                get_wheel_circumference_from_rig("2P"),
-                                bin_size,
-                                start,
-                                end,
-                                do_shuffle=True,
-                            )
-                            for trial in session.trials
-                            if trial.texture_rewarded and trial_is_imaged(trial)
-                        ]
-                    ),
-                    0,
-                ),
-                sigma=sigma_bins,
-                axis=1,
-            )
-            for _ in range(2000)
-        ]
-    )
-
-    place_threshold = np.percentile(shuffled_matrices, 99, axis=0)
-
-    shuffled_place_cells = np.array(
-        [
-            has_five_consecutive_trues(shuffled_matrices[idx, :, :] > place_threshold)
-            for idx in range(shuffled_matrices.shape[0])
-        ]
-    )
-
-    pcs = has_five_consecutive_trues(smoothed_matrix > place_threshold)
-
-    print(
-        f"percent place cells shuffed {np.mean(np.sum(shuffled_place_cells, axis=1) / shuffled_place_cells.shape[1])}"
-    )
-
-    print(f"percent place cells not shuffled {np.sum(pcs) / len(pcs)}")
-
-    plt.figure()
-    plt.imshow(
-        zscore(sort_matrix_peak(smoothed_matrix[pcs, :]), axis=1),
-        aspect="auto",
-        cmap="viridis",
-        vmin=0,
-        vmax=1,
-    )
-
-    plt.title("real")
-
-    plt.colorbar()
-    plt.legend()
-
-    plt.figure()
-
-    plt.imshow(
-        zscore(
-            sort_matrix_peak(shuffled_matrices[0, shuffled_place_cells[0, :], :]),
-            axis=1,
-        ),
-        aspect="auto",
-        cmap="viridis",
-        vmin=0,
-        vmax=1,
-    )
-
-    plt.title("shuffled")
-    plt.colorbar()
-    plt.legend()
-    plt.show()
-
-    return np.ndarray([])
-
-
 def place_cells(
     session: Cached2pSession, dff: np.ndarray, spks: np.ndarray, denoised: np.ndarray
 ) -> None:
 
     # TODO: Don't reuse this function obviously
+
     spks = binarise_spikes(spks)
     grosmark_place_field(session, spks)
     return
@@ -644,7 +477,9 @@ if __name__ == "__main__":
         < dff.shape[1]
     ), "Tiff is too short"
 
-    if "unsupervised" in session.session_type.lower():
-        place_cells_unsupervised(session, dff)
-    else:
-        place_cells(session, dff, spks, denoised)
+    do_classify(session, spks)
+
+    # if "unsupervised" in session.session_type.lower():
+    #     place_cells_unsupervised(session, dff)
+    # else:
+    #     place_cells(session, dff, spks, denoised)
